Fix #19524: Incorrect staff count beyond 32767 (#19530)

+++ src/openrct2-ui/windows/Park.cpp
@@ -190,8 +190,8 @@ static constexpr const WindowParkAward _parkAwards[] = {
 
 class ParkWindow final : public Window
 {
-    int16_t _numberOfStaff = -1;
-    int16_t _numberOfRides = -1;
+    int32_t _numberOfStaff = -1;
+    int32_t _numberOfRides = -1;
     uint8_t _peepAnimationFrame = 0;
 
 public:
